+++ ts/examples/tutorial/apps/server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @tutorial-example/server
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [cd0e937]
+- Updated dependencies [4b31e78]
+- Updated dependencies [2794296]
+  - @channel.io/app-sdk-core@0.22.0
+  - @channel.io/app-sdk-server@0.22.0
+
 ## 0.0.10
 
 ### Patch Changes